+++ src/lib/ngx-hm-carousel-dynamic.directive.ts
@@ -1,5 +1,7 @@
 import { Directive, Input, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
 
+// TODO: ssr problem should not hide on ssr
+// TODO: show number change should recaculate is show and init show number
 @Directive({
   selector: '[ngxHmCarouselDynamic]'
 })
@@ -7,20 +9,19 @@ export class NgxHmCarouselDynamicDirective implements OnInit {
 
   @Input('ngxHmCarouselDynamic') index: number;
   @Input('ngxHmCarouselDynamicLength') length: number;
+  @Input('ngxHmCarouselDynamicShow') show = 1;
   @Input('ngxHmCarouselDynamicIndex')
-
   set currentI(value: number) {
     if (!this.compelete) {
-
-      const nextI = value + 1;
-      const prevI = value - 1;
-
+      const nextI = value + this.show;
+      const prevI = value - this.show;
       if (
         this.index === 0 ||
         this.index === this.length - 1 ||
         this.index === nextI ||
         this.index === prevI ||
-        this.index === value) {
+        this.index === value
+      ) {
         this._view.createEmbeddedView(this._template);
         this.compelete = true;
       }
@@ -28,6 +29,7 @@ export class NgxHmCarouselDynamicDirective implements OnInit {
   }
 
   private compelete = false;
+
   constructor(
     private _view: ViewContainerRef,
     private _template: TemplateRef<any>) {